backend/requirements.txt
Flask==3.1.1
gunicorn==23.0.0
gspread==6.1.3
google-auth==2.40.1
google-api-python-client==2.100.0
schedule==1.2.1
python-dotenv==1.0.1